+++ tests/RedisSortedSetsTest.php
@@ -21,48 +21,12 @@ class RedisSortedSetsTest extends TestCase
         $this->keyOptional = 'SortedSets:Optional';
     }
 
-
-    /*
-     * ========================================================================
-     * zAdd
-     * ========================================================================
-     */
-
-
     /** @test */
-    public function redis_sorted_sets_zadd_multiple()
+    public function redis_sorted_sets_zadd()
     {
-        // Start from scratch
         $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        // T E S T  -----------------------------------------------------------
         $this->assertEquals(1, $this->redis->zAdd($this->key, 1.0, 'A'));
         $this->assertEquals(1, $this->redis->zAdd($this->key, 2.0, 'B'));
-        // T E S T  -----------------------------------------------------------
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zadd_duplicate()
-    {
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.0, 'A'));
-        $this->assertEquals(0, $this->redis->zAdd($this->key, 2.0, 'A'));
-        // T E S T  -----------------------------------------------------------
-        // Remove all the keys used
-        $this->assertEquals(1, $this->redis->delete($this->key));
-    }
-
-    /** @test */
-    public function redis_sorted_sets_zadd_single()
-    {
-        // Start from scratch
-        $this->assertGreaterThanOrEqual(0, $this->redis->delete($this->key));
-        // T E S T  -----------------------------------------------------------
-        $this->assertEquals(1, $this->redis->zAdd($this->key, 1.0, 'A'));
-        // Remove all the keys used
         $this->assertEquals(1, $this->redis->delete($this->key));
     }
 }
